requirements.txt
tensorflow>=2.8
scikit-learn>=1.0
pandas>=1.3
streamlit>=1.0
matplotlib>=3.5
Pillow>=9.0
joblib>=1.1